+++ test/src/test_cast.cpp
@@ -17,7 +17,7 @@ TEST(test_cast, api)
     std::vector<uint32_t> buffer = {1337, 110066};
 
     uint32_t* data = buffer.data();
-    uint32_t size = buffer.size() * sizeof(uint32_t);
+    uint64_t size = buffer.size() * sizeof(uint32_t);
 
     auto storage = storage::storage(data, size);
     EXPECT_EQ(storage.size(), size);
